+++ rs/nns/governance/src/timer_tasks/mod.rs
@@ -6,6 +6,7 @@ use crate::canister_state::GOVERNANCE;
 use crate::timer_tasks::calculate_distributable_rewards::CalculateDistributableRewardsTask;
 
 mod calculate_distributable_rewards;
+mod distribute_rewards;
 mod seeding;
 
 thread_local! {
@@ -15,4 +16,9 @@ thread_local! {
 pub fn schedule_tasks() {
     SeedingTask::new(&GOVERNANCE).schedule(&METRICS_REGISTRY);
     CalculateDistributableRewardsTask::new(&GOVERNANCE).schedule(&METRICS_REGISTRY);
+    run_distribute_rewards_periodic_task();
+}
+
+pub fn run_distribute_rewards_periodic_task() {
+    distribute_rewards::run_distribute_rewards_periodic_task(&GOVERNANCE, &METRICS_REGISTRY);
 }
